connect express to frontend with OtelData. Adjusted typings in chrome extension
