+++ chrome-extension/src/types/types.ts
@@ -1,0 +1,14 @@
+export interface OtelData {
+    traceId?: string,
+    spanId?: string,
+    applicationType?: string, 
+    originatingService?: string, 
+    method?: string, 
+    status?: number, 
+    protocol?: string, 
+    startTime: number,
+    endTime?: number,
+    size?: number,
+    type?: string, 
+    urlEndpoint?: string
+}
