+++ npm-package/src/controllers/streamController.ts
@@ -1,11 +1,15 @@
 import { Request, Response, NextFunction } from 'express';
 import { EventEmitter } from 'events';
+import { OtelData } from '../types/types';
 
 const otelEventEmitter = new EventEmitter();
 
 export const streamController = {
     emitEvent: (req: Request, res: Response, next: NextFunction) => {
-        otelEventEmitter.emit('newOtelEvent', 'hello');
+        // res.locals.telemetryData.forEach((metric: OtelData) => {
+        //     otelEventEmitter.emit('newOtelEvent', metric);
+        // })
+        otelEventEmitter.emit('newOtelEvent', res.locals.telemetryData);
         return next();
     },
 
@@ -15,11 +19,8 @@ export const streamController = {
         res.setHeader('Connection', 'keep-alive');
         res.setHeader('Access-Control-Allow-Origin', '*');
 
-        otelEventEmitter.on('newOtelEvent', (data) => {
+        otelEventEmitter.on('newOtelEvent', (data: OtelData) => {
             res.write('data: ' + `${data}\n\n`)
         })
-        // res.locals.metrics.forEach(((metric: any) => {
-        //     res.write('data: ' + `${metric}\n\n`)
-        // }))
     }
 }
