+++ npm-package/src/routers/streamRouter.ts
@@ -1,3 +1,4 @@
+import { otelController } from "../controllers/otelController";
 import { streamController } from "../controllers/streamController";
 import { Router, Request, Response } from 'express';
 
@@ -6,7 +7,7 @@ const streamRouter = Router();
 streamRouter.get('/sse', streamController.sendEvent, (req: Request, res: Response) => {
 })
 
-streamRouter.get('/test', streamController.emitEvent, (req: Request, res: Response) => {
+streamRouter.get('/otel', otelController.parseAllRequest, streamController.emitEvent, (req: Request, res: Response) => {
     res.json('completed')
 })
