+++ npm-package/src/server.ts
@@ -1,21 +1,16 @@
 import express, { Express, NextFunction, Request, Response, ErrorRequestHandler } from 'express';
 import cors from 'cors';
-// import { otelController } from './controllers/otelController';
-import { DefaultError } from '../types/server'
+import { DefaultError } from './types/types'
 import streamRouter from './routers/streamRouter';
-const PORT: number = 3002;
 
+const PORT: number = 3002;
 const app: Express = express();
 
 app.use(cors());
 app.use(express.json());
-
 app.use(express.urlencoded({extended: true}));
+// connect stream router
 app.use('/stream', streamRouter)
-// app.use('/', otelController.parseAllRequest, (req, res) => {
-//     //set up sending data to the front end here
-//     res.sendStatus(200);
-// })
 
 // express general error handler
 app.use((err: ErrorRequestHandler, req: Request, res: Response, next: NextFunction) => {
@@ -29,7 +24,7 @@ app.use((err: ErrorRequestHandler, req: Request, res: Response, next: NextFuncti
     console.log(err);
     return res.status(newErr.code).json(newErr.message);
 });
-
+// connect to express port
 app.listen(PORT, (): void => {
     console.log('NextInspect express npm package running on on port:' + PORT)
 });
