--- npm-package/types/server.d.ts
@@ -1,5 +1,0 @@
-export interface DefaultError {
-    log: string;
-    code: number;
-    message: string;
-}
